Interval32Box: Fix upperBoundAssign
There was a bug in `upperBoundAssign` such that when attempting to
assign a bounded box with an unbounded box, e.g., top(⟙), the resulting
box would *not* be top(⟙).

As a result of fixing this, analysis of methods with loops properly
merge paths.

This was partially discovered as a result of 8556277 ("Interval32Box:
fix overzealous `isTop` predicate", Tue May 11 10:57:12 2021 -0600).

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>