Merge only values from feasible paths
While IntervalBoxState appropriately selects the smallest enclosing
interval when asked to merge two values, it does not understand that
other values in the paths may be infeasible and therefore all of the
values in one box are infeasible and should not be merged together.  Add
checks to forward flow merge such that we only merge feasible paths.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>