Refactor DBS Internal Graph Representation
Some memory bugs appeared have caused several problems.  Notably, due to
how the first iteration used `Constraint` types for edges in the graph.
Since constraints are not unique in each graph or state, querying for
them can be problematic.  Furthermore, overriding `equals` means we need
to override `hashcode` otherwise, we can't reliably find the
constraint.  However, the implementation of hashcode is not stable.
Ultimately, by refactoring the graph internals away from the graph
state, we resolve the memory address issues of using JGraphT and we have
the additional bonus of starting the work for using a matrix
representation if so desired.

[0]: https://jgrapht.org/guide/VertexAndEdgeTypes

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>