Insert negated variables in canonical form
Previously, when handling negation expressions such as `lVar = neg c`,
the inserted encoding becomes $Z0 - lVar = c$.  This causes some
interesting consequences.  Specifically, the analysis starts
"discovering" relationships between variables that do not actually
exist.  Consider, for example, the program added as a test case for this
change.  The old behaviour generates a DBS state as follows:

```dot
b0 --- (0, =)   ---> Z0
b1 --- (50, =)  ---> Z0 ---- (60, =) ---> $b25
b2 --- (60, =)  ---> Z0
```

Since transitive edges must exist, this results in "strange"
relationships for `b0`, `b1`, and `b2` to relate to `$b25`.  The latter,
`b2`, being particularly interesting seeing as that relationship is not
a relationship the DBS is not expressive enough to capture in general.

By cannonicalizing the expression to zero in the form of `lVar - Z0 =
-C`, the analysis no longer learns of strange variable relationships.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>