Constraint: fix improper constraint ordering
Specifically when minimizing constraints, the constraints were
deconstructed and recombined using the minimum elements of the
Constraint tuple.  This, however, is clearly incorrect as noted in the
test cases in DBS where the output constraint after narrowing was `(= l0
2)` instead of the proper constraint `(<= l0 2)`.  The `max` function
did not have this issue as it directly used the `compareTo` ordering,
which while slightly incorrect, `max` would copy the constraint
wholesale instead of reconstructing a constraint as done in `min`.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>