intervals: tests: remove regex parsing
Since 1d7d7ee ("interval state: always print in local sorted order", Thu
Sep 16 23:01:42 2021 -0600), it's not longer necessary to parse the
output of the state map object to reconstruct the values of the state in
the output.  This was particularly fragile and with the introduction of
unbounded intervals, it was not robust.

Instead, simply perform direct string comparison.  The sorted order is
stable and based on the lexigraphic ordering of the string
representation of Locals.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>