DBM: refactor assignments through setConstraint
Because we want to de-duplicate any extra logic that is applied when we
assign a constraint, refactor all current assignments of constraints to
use `setConstraint`.  As an added benefit, the logic for setConstraint
makes full closures idempotent in all cases.  That is, before, when the
closure operation executes against an infeasible matrix, repeated
closure applications would change the underlying values of the matrix.
Now, everything is assigned to bottom (⟘) (since it's the most minimal
element of the domain) and repeated applications of closure to a
"bottomed" matrix maintains bottom.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>