Soot: split initialization from method retrieval
Previously, we initialize (set the classpath) Soot and retrieve the
method we wish to analyze.  However, this does not work well when
attempting to analyze several methods in the same JVM instance.  This
refactoring changes how soot is initialized, providing a mechanism for
loading multiple classes at initialization, and how analysis runners
retrieve the methods and subsequently the active body for each method.

This change is mostly cosmetic to the existing analyzers.  Analyzers
attempting to compare two methods, however, need this change to
function.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>