Create new GraphProjection object for DOT exports
The reason behind DBSConstraint was to act as a proper "edge" for
JGraphT according to [0].  That is, the default `ZoneConstraint`
overrides `equals` and `hashCode` making two `ZoneContraint` objects
equal if they contain the same integer value.  However, this breaks the
representation in JGraphT.  Therefore, a similar class was created which
does not override these functions.  But since we want to get rid of the
two constraint classes to simplify the API, we need a new way of
capturing the graph projection.  To obscure the graph details from each
of the state objects, we create a `GraphProjection` class.  This class
encapsulates the details of exporting to DOT format and representing
each numerical abstract state as a "graph".

This sets the path for renaming `ZoneConstraint` to simply `Constraint`.

[0]: https://jgrapht.org/guide/VertexAndEdgeTypes

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>