Unify SMT report outputs across analyses
Instead of having two different output formats for the different report
"types", unify them into a single format.  This means that there only
needs to be one report "parser".  The difference between the report
types comes down to the formatter.  The "full" and "changed" formatters
are essentially the same now, the consume a single file for each domain
and output the line from the respective files.  The "min" and
"reachable" formatters (reachable doesn't exist in this commit) consume
both the full output and the "changed" output and selectively output the
union of both seen identifiers from either side.  In this way, the "min"
and the "reachable" formatter can handle relational domains with more
precision.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>