add "reachable" state output
There are a couple of different output versions we're working with at
the moment.  One, where we do not consider direction when getting a
changed subset of the state from zones.  This new one considers
direction and strictly enforces "reachibility" in the output SMT
expression.  Unsurprisingly, "reachable" and "connected" are the same
for intervals and predicates.  However, to keep things "easy", we have
all domains output the additional "reachable" formulas.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>