ZoneStates: use inState when projecting to intervals
This fixes a bug where Zones was computing the wrong interval when
unable to directly encode relational constraints.

However, this may never have been an issue in practice seeing as the
"integration" tests succeeded.  Due to some other changes, the test
suite's setup for some of the internal methods uncovered the
inconsistency.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>