SmtExpression: Expand types of BinopSmtExpressions
Instead of pushing the Grimp expression directly into the
BinopSmtExpression types, push the left/right SmtExpressions and expand
the types to properly extract the appropriate value.  This should push
appropriate decision questions into the appropriate layer so that only a
single path of code needs to exist at the right layer.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>