SmtExpression: rename and add new getValue
The behvaiour of `getValue` is to return the connected set of values, so
any expression which contains any variable of the source set is
returned.  However, this really should be named `getConnectedValue`.
Instead, `getValue` should return only expressions which contain a
combination of the source variables.  That is, binary operator
expressions should only return when some pair of variables are on both
sides.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>