Smt2Format: Allow specification of SMT Logic
At the top of the entailed file, we set a default LIA logic.  However,
since we might want to adjust this, parameterize the formatter to set
different logics.  Currently, only the following sequence of logics are
supported.

QF_LIA → LIA → NIA → UFNIA → AUFNIA.

If no logic is specified, the formatter defaults to LIA.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>