Construct SMT expressions without use of solver
Do not use `SolverWrapper` to format SmtExpressions as SMTLIB
expressions.  There's a lot of complexity in converting to BinopExprs
and then into Z3 Context objects, only to call `.toString` on the
result.  Instead, we eseentially have the information already embedded
in the SmtExpression type tree, lean on this to construct the
appropriate expressions.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>