smt/Number: use longs for numerical values
The minimum integer is larger than the max integer if the sign is
removed, which results in an exception when attempting to parse.
Perhaps not the most efficient solution, since we know the actual output
is always within the bounds of an integer, we can simply use the long
type instead.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>