Pass parameters to analysis using an options object
Instead of increasing the number of parameters passed for each type of
numerical analysis, create an options context object that can be
constructed from user options and parameters, and pass the result.

This had an unfortunate consequence of shoveling a bunch of snow over
the parameterized types for StateFactory causing a whole mess.  Thus,
there were some redesigns of the IntegerAnalysis surface such that we
can have an "Options" object without type parameters.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>